feat: use loader on each section on admin
